extract Create and Edit into ProjectForm
